Create CJS export for allure reporter
