+++ e2e/interop/allure-reporter.e2e.js
@@ -1,0 +1,19 @@
+const assert = require('node:assert')
+const AllureReporter = require('../../packages/wdio-allure-reporter')
+const { addFeature } = require('../../packages/wdio-allure-reporter')
+
+console.log('Test Allure Reporter Exports')
+assert.equal(typeof addFeature, 'function')
+assert.equal(typeof AllureReporter.addLabel, 'function')
+assert.equal(typeof AllureReporter.addSeverity, 'function')
+assert.equal(typeof AllureReporter.addIssue, 'function')
+assert.equal(typeof AllureReporter.addTestId, 'function')
+assert.equal(typeof AllureReporter.addStory, 'function')
+assert.equal(typeof AllureReporter.addEnvironment, 'function')
+assert.equal(typeof AllureReporter.addDescription, 'function')
+assert.equal(typeof AllureReporter.addAttachment, 'function')
+assert.equal(typeof AllureReporter.startStep, 'function')
+assert.equal(typeof AllureReporter.endStep, 'function')
+assert.equal(typeof AllureReporter.addStep, 'function')
+assert.equal(typeof AllureReporter.addArgument, 'function')
+console.log('Allure Reporter CJS Test Passed!')
