+++ packages/wdio-allure-reporter/cjs/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "commonjs"
+}
