+++ packages/wdio-allure-reporter/tests/api.test.ts
@@ -1,107 +1,135 @@
 import path from 'node:path'
-import { describe, it, expect, beforeEach, vi } from 'vitest'
+import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
 
-import reporter from '../src/index.js'
-import * as utils from '../src/utils.js'
+import reporter, {
+    addFeature, addLabel, addSeverity, addIssue, addTestId, addStory,
+    addEnvironment, addDescription, addAttachment, startStep, endStep,
+    addStep, addArgument
+} from '../src/index.js'
 import { events, stepStatuses } from '../src/constants.js'
 
 vi.mock('@wdio/reporter', () => import(path.join(process.cwd(), '__mocks__', '@wdio/reporter')))
 vi.mock('../src/utils')
 
 describe('reporter reporter api', () => {
+    const processEmit = process.emit.bind(process)
+    beforeEach(() => {
+        process.emit = vi.fn() as any
+    })
+
+    afterEach(() => {
+        process.emit = processEmit
+    })
+
+    it('exports correct API', () => {
+        expect(typeof addFeature).toBe('function')
+        expect(typeof addLabel).toBe('function')
+        expect(typeof addSeverity).toBe('function')
+        expect(typeof addIssue).toBe('function')
+        expect(typeof addTestId).toBe('function')
+        expect(typeof addStory).toBe('function')
+        expect(typeof addEnvironment).toBe('function')
+        expect(typeof addDescription).toBe('function')
+        expect(typeof addAttachment).toBe('function')
+        expect(typeof startStep).toBe('function')
+        expect(typeof endStep).toBe('function')
+        expect(typeof addStep).toBe('function')
+        expect(typeof addArgument).toBe('function')
+    })
+
     it('should pass correct data from addLabel', () => {
         reporter.addLabel('customLabel', 'Label')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addLabel, { name: 'customLabel', value: 'Label' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addLabel, { name: 'customLabel', value: 'Label' })
     })
 
     it('should pass correct data from addStory', () => {
         reporter.addStory('Story')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addStory, { storyName: 'Story' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addStory, { storyName: 'Story' })
     })
 
     it('should pass correct data from addFeature', () => {
         reporter.addFeature('foo')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addFeature, { featureName: 'foo' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addFeature, { featureName: 'foo' })
     })
 
     it('should pass correct data from addSeverity', () => {
         reporter.addSeverity('foo')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addSeverity, { severity: 'foo' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addSeverity, { severity: 'foo' })
     })
 
     it('should pass correct data from addIssue', () => {
         reporter.addIssue('1')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addIssue, { issue: '1' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addIssue, { issue: '1' })
     })
 
     it('should pass correct data from addTestId', () => {
         reporter.addTestId('2')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addTestId, { testId: '2' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addTestId, { testId: '2' })
     })
 
     it('should pass correct data from addEnvironment', () => {
         reporter.addEnvironment('foo', 'bar')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addEnvironment, { name: 'foo', value: 'bar' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addEnvironment, { name: 'foo', value: 'bar' })
     })
 
     it('should pass correct data from addDescription', () => {
         reporter.addDescription('foo', 'html')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addDescription, { description: 'foo', descriptionType: 'html' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addDescription, { description: 'foo', descriptionType: 'html' })
     })
 
     it('should pass correct data from addStep', () => {
         reporter.addStep('foo', { name: 'bar', content: 'baz', type: 'text/plain' }, stepStatuses.FAILED )
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledTimes(1)
         const step = { 'step': { 'attachment': { 'content': 'baz', 'name': 'bar', 'type': 'text/plain' }, 'status': 'failed', 'title': 'foo' } }
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addStep, step)
+        expect(process.emit).toHaveBeenCalledWith(events.addStep, step)
     })
 
     it('should support default attachment name for addStep', () => {
         reporter.addStep('foo', { content: 'baz' }, stepStatuses.FAILED )
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledTimes(1)
         const step = { 'step': { 'attachment': { 'content': 'baz', 'name': 'attachment', 'type': 'text/plain' }, 'status': 'failed', 'title': 'foo' } }
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addStep, step)
+        expect(process.emit).toHaveBeenCalledWith(events.addStep, step)
     })
 
     it('should support default attachment type for addStep', () => {
         reporter.addStep('foo', { content: 'baz' })
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledTimes(1)
         const step = { 'step': { 'attachment': { 'content': 'baz', 'name': 'attachment', 'type': 'text/plain' }, 'status': 'passed', 'title': 'foo' } }
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addStep, step)
+        expect(process.emit).toHaveBeenCalledWith(events.addStep, step)
     })
 
     it('should support start step', () => {
         reporter.startStep('foo')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.startStep, 'foo')
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.startStep, 'foo')
     })
 
     it('should support end step', () => {
         reporter.endStep('passed')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.endStep, 'passed')
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.endStep, 'passed')
     })
 
     it('should support addStep without attachment', () => {
         reporter.addStep('foo')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledTimes(1)
         const step = { 'step': { 'status': 'passed', 'title': 'foo' } }
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addStep, step)
+        expect(process.emit).toHaveBeenCalledWith(events.addStep, step)
     })
 
     it('should support default step status for addStep', () => {
         reporter.addStep('foo', { content: 'baz' } )
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledTimes(1)
         const step = { 'step': { 'attachment': { 'content': 'baz', 'name': 'attachment', 'type': 'text/plain' }, 'status': 'passed', 'title': 'foo' } }
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addStep, step)
+        expect(process.emit).toHaveBeenCalledWith(events.addStep, step)
     })
 
     it('should throw exception for incorrect status for addStep', () => {
@@ -118,8 +146,8 @@ describe('reporter reporter api', () => {
 
     it('should pass correct data from addArgument', () => {
         reporter.addArgument('os', 'osx')
-        expect(utils.tellReporter).toHaveBeenCalledTimes(1)
-        expect(utils.tellReporter).toHaveBeenCalledWith(events.addArgument, { name: 'os', value: 'osx' })
+        expect(process.emit).toHaveBeenCalledTimes(1)
+        expect(process.emit).toHaveBeenCalledWith(events.addArgument, { name: 'os', value: 'osx' })
     })
 
     describe('addAttachment', () => {
@@ -152,7 +180,7 @@ describe('reporter reporter api', () => {
         scenarios.forEach(({ title, actualType, content, name, type }) => {
             it(title, () => {
                 reporter.addAttachment(name, content, actualType!)
-                expect(utils.tellReporter).toBeCalledWith(events.addAttachment, { name, content, type })
+                expect(process.emit).toBeCalledWith(events.addAttachment, { name, content, type })
             })
         })
     })
