+++ packages/wdio-allure-reporter/tsconfig.cjs.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.prod.json",
+  "compilerOptions": {
+    "module": "CommonJS",
+    "outDir": "./cjs",
+    "declaration": false,
+    "declarationMap": false
+  },
+  "include": [
+    "src/common"
+  ]
+}
